requirements.txt
Django==2.0.3
django-cors-headers==2.2.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.7.7
mongoengine==0.15.0
pymongo==3.6.1
pytz==2018.3
six==1.11.0
